type tweak for windows build
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@6055 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/applications/mod_voicemail/mod_voicemail.c
@@ -684,7 +684,7 @@ static switch_status_t vm_macro_get(switch_core_session_t *session,
         memset(buf, 0, buflen);
         args.input_callback = cancel_on_dtmf;
         args.buf = buf;
-        args.buflen = buflen;
+        args.buflen = (uint32_t)buflen;
         ap = &args;
     }
